Remove Popen.wait() call
Using Popen.wait() is incorrect usage here.  I should have reviewed this change more carefully.

Popen.communicate() already waits for the process to terminate, documented here:
https://docs.python.org/2/library/subprocess.html#subprocess.Popen.communicate

In addition, the usage of Popen.wait() when using PIPE can lead to deadlocks, with the recommendation to use Popen.communicate() which was already present, documented here:
https://docs.python.org/2/library/subprocess.html#subprocess.Popen.wait

The correct call would be to retrieve the already set returncode value after Popen.communicate() has returned.
https://docs.python.org/2/library/subprocess.html#subprocess.Popen.returncode

This is actually a bugfix that needs to propagate anywhere this code was deployed to avoid hanging processes on upload.