Bugfixes to setup.py, convert taskrunner
opt.pl is in two directories, which makes setup.py fail because it tries
to link the file twice with the same link location. This makes
os.symlink fail. Added code to print a warning instead.

Made the following fixes to the convert taskrunner:
- removed debug mode: debug mode forced the script to fail because it
skipped decoding optional_arguments and job_details from base64 encoded
JSON. Later in the script job_details is treated as a dict.
- made job_details required, because it is. If it is not provided, the
script will fail.
- clarified that job_details and optional_arguments are base64 encoded
JSON.
- correctly handled the case where optional_arguments are not provided.
The previous behavior would set it to a dict, and then try to base64
unencode the dict, which would fail.